+++ src/main/java/com/nhnacademy/minidoorayprojectapi/domain/task/service/TaskService.java
@@ -1,5 +1,8 @@
 package com.nhnacademy.minidoorayprojectapi.domain.task.service;
 
+import com.nhnacademy.minidoorayprojectapi.domain.comment.dao.CommentRepository;
+import com.nhnacademy.minidoorayprojectapi.domain.comment.dto.response.CommentDto;
+import com.nhnacademy.minidoorayprojectapi.domain.comment.entity.Comment;
 import com.nhnacademy.minidoorayprojectapi.domain.milestone.dao.MilestoneRepository;
 import com.nhnacademy.minidoorayprojectapi.domain.milestone.dto.response.MilestoneDto;
 import com.nhnacademy.minidoorayprojectapi.domain.milestone.exception.MilestoneNotFoundException;
@@ -19,6 +22,7 @@ import com.nhnacademy.minidoorayprojectapi.domain.task.dto.response.TaskSeqNameA
 import com.nhnacademy.minidoorayprojectapi.domain.task.entity.Task;
 import com.nhnacademy.minidoorayprojectapi.domain.task.entity.TaskTag;
 import com.nhnacademy.minidoorayprojectapi.domain.task.exception.TaskNotFoundException;
+import com.nhnacademy.minidoorayprojectapi.global.exception.UnauthorizedAccessException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageImpl;
@@ -39,6 +43,7 @@ public class TaskService {
     private final MilestoneRepository milestoneRepository;
     private final TagRepository tagRepository;
     private final TaskTagRepository taskTagRepository;
+    private final CommentRepository commentRepository;
 
 
     /**
@@ -61,14 +66,15 @@ public class TaskService {
         return new PageImpl<>(taskPageDto,taskPage.getPageable(),taskPage.getTotalElements());
     }
 
-    public TaskDto getTask(Long projectSeq, Long taskSeq){
+    public TaskDto getTask(Long projectSeq, Long taskSeq, Pageable pageable){
         Task task = taskRepository.findByProject_ProjectSeqAndTaskSeq(projectSeq, taskSeq)
                 .orElseThrow(TaskNotFoundException::new);
         List<Tag> tags = taskTagRepository.findAllByTask_TaskSeq(taskSeq);
-        return convertToTaskDto(task, tags);
+        Page<Comment> comments = commentRepository.getAllByTask_TaskSeq(taskSeq,pageable);
+        return convertToTaskDto(task, tags, comments);
     }
 
-    private TaskDto convertToTaskDto(Task task, List<Tag> tags){
+    private TaskDto convertToTaskDto(Task task, List<Tag> tags, Page<Comment> comments){
         return TaskDto.builder()
                 .taskSeq(task.getTaskSeq())
                 .taskTitle(task.getTaskTitle())
@@ -92,6 +98,17 @@ public class TaskService {
                                         .tagName(tag.getTagName())
                                         .build())
                                 .collect(Collectors.toList())
+                )
+                .comments(
+                        new PageImpl<>(task.getComments().stream()
+                                .map(comment -> new CommentDto(comment.getCommentSeq(),
+                                        comment.getMemberSeq(),
+                                        comment.getCommentContent(),
+                                        comment.getCommentCreatedAt()))
+                                .collect(Collectors.toList()),
+                                comments.getPageable(),
+                                comments.getTotalElements()
+                        )
                 )
                 .build();
     }
@@ -129,20 +146,13 @@ public class TaskService {
         return convertToTaskSeqDto(newTask);
     }
 
-//    public void createTaskTag(Long projectSeq, Long taskSeq, Task task){
-//        Tag tag = tagRepository.findByProject_ProjectSeqAndTagSeq(projectSeq,taskSeq)
-//                .orElseThrow(TagNotFoundException::new);
-//        TaskTag.builder()
-//                .taskTagPk(new TaskTag.TaskTagPk(task.getTaskSeq(), tag.getTagSeq()))
-//                .tag(tag)
-//                .task(task)
-//                .build();
-//    }
-
     @Transactional
     public TaskSeqDto updateTask(Long projectSeq, Long taskSeq, TaskUpdateRequestDto taskUpdateRequest){
         Task task = taskRepository.findByProject_ProjectSeqAndTaskSeq(projectSeq, taskSeq)
                 .orElseThrow(TaskNotFoundException::new);
+        if(!task.getMemberSeq().equals(task.getMemberSeq())){
+            throw new UnauthorizedAccessException("접근 권한이 없습니다.");
+        }
         task.updateTask(taskUpdateRequest.getTaskTitle(),taskUpdateRequest.getTaskContent(),
                 taskUpdateRequest.getTaskStatus(),
                 milestoneRepository.findByProject_ProjectSeqAndMilestoneSeq(projectSeq,
@@ -157,7 +167,10 @@ public class TaskService {
 
     //TODO task 삭제할 때 tag도 함께 삭제해야하나? 다대다??
     @Transactional
-    public void deleteTask(Long taskSeq){
+    public void deleteTask(Long taskSeq, Long memberSeq){
+        if(taskRepository.existsByTaskSeqAndAndMemberSeq(taskSeq,memberSeq)){
+            throw new UnauthorizedAccessException("접근 권한이 없습니다.");
+        }
         taskRepository.deleteById(taskSeq);
     }
